feat: move TalkSession to session module. Create stream module
